Friends of GNOME: Explained the "no gift" option also means no LWN subscription
This change indicates a slightly modified policy for the LWN tokens.
The intention is to allow the donor to "fully" donate the money.
We might want to allow for no gift, but LWN subscription in the future.
For now, it would probably mean to introduce another stereobox which is
nothing I am keen on doing right now.
Donors who want no gift to be shipped but want to receive an LWN token
should probably get in touch via email.