Groupon: Updated the *Template*
You must not edit the .html, but the template. Then run make and the
base directory of the repository.  This wouldn't have happened if we
still had server side builds.  Sorry for the confusion. The automatic
commit of the amonut of the money raised was barfing over a rebase
conflict, hence there were no update for the last few hours.